pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-deblurring"
version = "0.1.0"
description = "Image deblurring training pipeline with Hydra and PyTorch Lightning"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Image Deblurring Team" },
]
dependencies = [
  "torch==2.3.1",
  "torchvision==0.18.1",
  "lightning",
  "numpy==2.1.2",
  "albucore==0.0.21",
  "albumentations==1.4.23",
  "opencv-python==4.10.0.84",
  "scikit-image==0.24.0",
  "scipy==1.14.1",
  "tqdm==4.66.5",
  "pandas==2.2.3",
  "pytorch-msssim==1.0.0",
  "einops==0.8.0",
  "hydra-core==1.3.2",
  "omegaconf==2.3.0",
  "mlflow==2.16.0",
  "matplotlib",
  "pillow",
  "dvc",
]

[project.optional-dependencies]
dev = [
  "pre-commit",
  "ruff",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["E203", "E501"]
